+++ Merger/package.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0"?>
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<package format="3">
+  <name>pcd_merge</name>
+  <version>0.0.0</version>
+  <description>Merges front, left, and right .bin files with corresponding timestamps</description>
+  <maintainer email="abauris@gmail.com">aurimasbalciunas</maintainer>
+  <license>TODO: License declaration</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+
+  <test_depend>ament_lint_auto</test_depend>
+  <test_depend>ament_lint_common</test_depend>
+
+  <depend>rclcpp</depend>
+  <depend>sensor_msgs</depend>
+  <depend>message_filters</depend>
+  <depend>Eigen3</depend>
+<export>
+    <build_type>ament_cmake</build_type>
+</export>
+
+</package>
